refactoring and allowing service schedulers to select permission groups like designated_service
