+++ ap/services/admin.py
@@ -19,20 +19,17 @@ from services.models import (
     ServiceException,
 )
 
+from .forms import (
+  SeasonalServiceScheduleForm,
+  AssignmentAdminForm,
+  WorkGroupAdminForm,
+  ExceptionAdminForm,
+  QualificationForm
+)
+
 from aputils.admin_utils import FilteredSelectMixin
 from aputils.widgets import MultipleSelectFullCalendar
 from aputils.queryfilter import QueryFilterService
-from aputils.custom_fields import CSIMultipleChoiceField
-
-
-# This is written to improve query performance on admin backend
-class WorkerPrejoinMixin(forms.ModelForm):
-  workers = forms.ModelMultipleChoiceField(
-      label='Workers',
-      queryset=Worker.objects.select_related('trainee').all(),
-      required=False,
-      widget=admin.widgets.FilteredSelectMultiple('workers', is_stacked=False)
-  )
 
 
 class ReadonlyException(object):
@@ -114,22 +111,6 @@ class WorkerAdmin(admin.ModelAdmin):
   class Meta:
     model = Worker
     fields = '__all__'
-
-
-class SeasonalServiceScheduleForm(forms.ModelForm):
-  services = forms.ModelMultipleChoiceField(
-      label='Services',
-      queryset=Service.objects.all(),
-      required=False,
-      widget=admin.widgets.FilteredSelectMultiple('services', is_stacked=False)
-  )
-
-  class Meta:
-    model = SeasonalServiceSchedule
-    exclude = []
-    widgets = {
-        'services': admin.widgets.FilteredSelectMultiple('services', is_stacked=False),
-    }
 
 
 class SeasonalServiceScheduleAdmin(FilteredSelectMixin, admin.ModelAdmin):
@@ -254,12 +235,6 @@ class ServiceAdmin(admin.ModelAdmin):
     fields = '__all__'
 
 
-class AssignmentAdminForm(WorkerPrejoinMixin, forms.ModelForm):
-  class Meta:
-    model = Assignment
-    fields = '__all__'
-
-
 class AssignmentAdmin(admin.ModelAdmin):
   form = AssignmentAdminForm
   list_display = ('week_schedule', 'service', 'service_slot', 'worker_list', 'workers_needed', 'pin')
@@ -274,14 +249,6 @@ class AssignmentAdmin(admin.ModelAdmin):
 
   class Meta:
     model = Assignment
-    fields = '__all__'
-
-
-class WorkGroupAdminForm(WorkerPrejoinMixin, forms.ModelForm):
-  query_filters = CSIMultipleChoiceField(choices=QueryFilterService.get_choices(), required=False, label='Filters')
-
-  class Meta:
-    model = WorkerGroup
     fields = '__all__'
 
 
@@ -321,20 +288,24 @@ class WorkerGroupAdmin(admin.ModelAdmin):
     model = WorkerGroup
     fields = '__all__'
 
+  # saves the ManyToMany relation of permission groups with trainees
+  def save_related(self, request, form, formsets, change):
+    super(WorkerGroupAdmin, self).save_related(request, form, formsets, change)
+    instance = form.save(commit=False)
+    permission_groups = instance.permission_groups.all()
+    workers = instance.workers.all()
+    trainees = Trainee.objects.filter(worker__in=workers)
+    for g in permission_groups:
+      for t in trainees:
+        # adds trainee to the permission group
+        g.user_set.add(t)
+      g.save()
+
 
 # method for updating
 @receiver(post_save, sender=Qualification)
 def add_query_filter(sender, instance, **kwargs):
   QueryFilterService.addQ(instance.name, worker__qualifications__name=instance.name)
-
-class ExceptionAdminForm(WorkerPrejoinMixin, forms.ModelForm):
-
-  class Meta:
-    model = ServiceException
-    fields = '__all__'
-    widgets = {
-        'services': MultipleSelectFullCalendar(Service.objects.all(), 'services'),
-    }
 
 
 class ExceptionAdmin(admin.ModelAdmin):
@@ -354,16 +325,6 @@ class ExceptionAdmin(admin.ModelAdmin):
   class Meta:
     model = ServiceException
     fields = '__all__'
-
-
-class QualificationForm(WorkerPrejoinMixin, forms.ModelForm):
-
-  class Meta:
-    model = Qualification
-    exclude = []
-    widgets = {
-        'workers': admin.widgets.FilteredSelectMultiple('workers', is_stacked=False),
-    }
 
 
 class QualificationAdmin(FilteredSelectMixin, admin.ModelAdmin):
@@ -443,19 +404,14 @@ class WeekScheduleAdmin(admin.ModelAdmin):
 
 
 admin.site.register(SeasonalServiceSchedule, SeasonalServiceScheduleAdmin)
-
 admin.site.register(Category, CategoryAdmin)
 admin.site.register(Service, ServiceAdmin)
 admin.site.register(ServiceSlot, ServiceSlotAdmin)
-
 admin.site.register(Qualification, QualificationAdmin)
 admin.site.register(Worker, WorkerAdmin)
-
 # admin.site.register(Worker, WorkerAdmin2)
 
 admin.site.register(WorkerGroup, WorkerGroupAdmin)
-
 admin.site.register(ServiceException, ExceptionAdmin)
-
 admin.site.register(Assignment, AssignmentAdmin)
 admin.site.register(WeekSchedule, WeekScheduleAdmin)
