+++ ap/services/models/workergroup.py
@@ -86,11 +86,7 @@ class WorkerGroup(models.Model):
 
   # Algorithm will assign higher priority first
   assign_priority = models.PositiveSmallIntegerField(default=1)
-  # permission_groups = models.ManyToManyField(Group, related_name='permission_groups', blank=True)
-  # permission_groups = models.TextField(blank=True, null=True)
   permission_groups = models.ManyToManyField(Group, related_name='service_group')
-  # permission_groups = ('Test', 'Test2', 'Test3') #models.TextField(blank=True)
-
   last_modified = models.DateTimeField(auto_now=True)
 
   @cached_property
@@ -155,18 +151,3 @@ class WorkerGroup(models.Model):
 @receiver(post_save, sender=Qualification)
 def add_query_filter(sender, instance, **kwargs):
   QueryFilterService.addQ(instance.name, worker__qualifications__name=instance.name)
-
-# method for updating permissions groups
-@receiver(post_save, sender=WorkerGroup)
-def add_workergroup_permissions(sender, instance, **kwargs):
-  permission_groups = instance.permission_groups.all()
-  workers = instance.workers.all()
-  # workers_sender = sender.trainees.all()
-  trainees = Trainee.objects.filter(worker__in=workers)
-  # print permission_groups, trainees
-  # print "sender's workers: ", workers_sender
-  # print "instance's workers: ", workers
-  for g in permission_groups:
-    for t in trainees:
-      g.user_set.add(t)
-    g.save()
